Improve init server to start as the first thing
