+++ backend/rootLog.ts
@@ -1,17 +1,9 @@
-import { config } from 'dotenv';
 import fs from 'node:fs';
 import path from 'path';
 import * as winston from 'winston';
 
+import '@backend/env';
 import 'winston-daily-rotate-file';
-
-// load .env and .env.local files. needed because this gets loaded before Next.js loads the .env files
-config({
-    path: [
-        path.resolve(process.cwd(), '.env'),
-        path.resolve(process.cwd(), '.env.local'),
-    ],
-});
 
 const logLevel = process.env.LOG_LEVEL?.toLowerCase();
